+++ src/index.d.ts
@@ -1,7 +1,12 @@
 import * as React from 'react';
 
+type CustomContext = React.Context<unknown>;
+
+export type createCustomContext = () => CustomContext;
+
 export type TrackedProviderProps<S, D> = {
   useValue: () => [S, D];
+  customContext?: CustomContext;
 };
 
 export type TrackedProviderType<S = unknown, D = unknown>
@@ -9,11 +14,16 @@ export type TrackedProviderType<S = unknown, D = unknown>
 
 export const TrackedProvider: TrackedProviderType;
 
-export const useTracked: <S, D>() => [S, D];
+type Opts = {
+  customContext?: CustomContext;
+};
 
-export const useDispatch: <D>() => D;
+export const useTracked: <S, D>(opts?: Opts) => [S, D];
+
+export const useDispatch: <D>(opts?: Opts) => D;
 
 export const useSelector: <S, M>(
   selector: (state: S) => M,
-  equalityFn?: (a: unknown, b: unknown) => boolean,
+  equalityFn?: (a: M, b: M) => boolean | Opts & { equalityFn?: (a: M, b: M) => boolean },
+  opts?: Opts,
 ) => M;